add test_rhythm_to_midi
